Draw areas in the 3D view and give them a psuedo position
