+++ src/editor/renderer.h
@@ -59,6 +59,7 @@ struct RenderSettings {
 	bool draw_sound_instances = true;
 	bool draw_paths = true;
 	bool draw_grind_paths = true;
+	bool draw_areas = true;
 	bool draw_collision = false;
 	
 	bool draw_selected_instance_normals = false;
